+++ pkg/integrations/elasticsearch_exporter/elasticsearch_exporter.go
@@ -4,13 +4,17 @@ package elasticsearch_exporter //nolint:golint
 
 import (
 	"context"
+	"encoding/base64"
 	"fmt"
 	"net/http"
 	"net/url"
+	"os"
+	"strings"
 	"time"
 
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
+	commoncfg "github.com/grafana/agent/component/common/config"
 	"github.com/grafana/agent/pkg/integrations"
 	integrations_v2 "github.com/grafana/agent/pkg/integrations/v2"
 	"github.com/grafana/agent/pkg/integrations/v2/metricsutils"
@@ -66,6 +70,19 @@ type Config struct {
 	ExportDataStreams bool `yaml:"data_stream,omitempty"`
 	// Export stats for Snapshot Lifecycle Management
 	ExportSLM bool `yaml:"slm,omitempty"`
+	// BasicAuth block allows secure connection with Elasticsearch cluster via Basic-Auth
+	BasicAuth *commoncfg.BasicAuth `yaml:"basic_auth,omitempty"`
+}
+
+// Custom http.Transport struct for Basic Auth-secured communication with ES cluster
+type BasicAuthHTTPTransport struct {
+	http.Transport
+	authHeader string
+}
+
+func (b *BasicAuthHTTPTransport) RoundTrip(req *http.Request) (*http.Response, error) {
+	req.Header.Add("authorization", b.authHeader)
+	return b.Transport.RoundTrip(req)
 }
 
 // UnmarshalYAML implements yaml.Unmarshaler for Config
@@ -115,12 +132,29 @@ func New(logger log.Logger, c *Config) (integrations.Integration, error) {
 	// returns nil if not provided and falls back to simple TCP.
 	tlsConfig := createTLSConfig(c.CA, c.ClientCert, c.ClientPrivateKey, c.InsecureSkipVerify)
 
-	httpClient := &http.Client{
-		Timeout: c.Timeout,
-		Transport: &http.Transport{
+	esHttpTransport := &BasicAuthHTTPTransport{
+		Transport: http.Transport{
 			TLSClientConfig: tlsConfig,
 			Proxy:           http.ProxyFromEnvironment,
 		},
+	}
+
+	if c.BasicAuth != nil {
+		password := string(c.BasicAuth.Password)
+		if len(c.BasicAuth.PasswordFile) > 0 {
+			buff, err := os.ReadFile(c.BasicAuth.PasswordFile)
+			if err != nil {
+				return nil, fmt.Errorf("unable to load password file %s: %w", c.BasicAuth.PasswordFile, err)
+			}
+			password = strings.TrimSpace(string(buff))
+		}
+		encodedAuth := base64.StdEncoding.EncodeToString([]byte(c.BasicAuth.Username + ":" + password))
+		esHttpTransport.authHeader = "Basic " + encodedAuth
+	}
+
+	httpClient := &http.Client{
+		Timeout:   c.Timeout,
+		Transport: esHttpTransport,
 	}
 
 	clusterInfoRetriever := clusterinfo.New(logger, httpClient, esURL, c.ExportClusterInfoInterval)